Commonize `ktor-server-double-receive`
