+++ ktor-server/ktor-server-plugins/ktor-server-double-receive/jvm/test/DoubleReceiveTestJvm.kt
@@ -1,6 +1,5 @@
 import io.ktor.server.plugins.doublereceive.*
 import io.ktor.utils.io.*
-import io.ktor.utils.io.core.*
 import kotlinx.coroutines.*
 import kotlinx.io.*
 import kotlin.coroutines.*
@@ -10,7 +9,7 @@ import kotlin.test.*
  * Copyright 2014-2022 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
  */
 
-class DoubleReceiveTest {
+class DoubleReceiveTestJvm {
 
     @Test
     fun testFileCache() = runBlocking {
@@ -23,21 +22,6 @@ class DoubleReceiveTest {
 
         repeat(3) {
             val received = cache.read().readRemaining().readByteArray()
-            assertEquals(content.toList(), received.toList())
-        }
-    }
-
-    @Test
-    fun testInMemoryCache() = runBlocking {
-        val content = ByteArray(16 * 1024 * 1024) { it.toByte() }
-        val cache = MemoryCache(
-            ByteReadChannel(content),
-            EmptyCoroutineContext
-        )
-
-        repeat(3) {
-            val channel = cache.read()
-            val received = channel.readRemaining().readByteArray()
             assertEquals(content.toList(), received.toList())
         }
     }
